+code simplifications
